リファクタリング？
* 艦種/装備種別を列挙型に変更
    * 実装上の問題というか互換性維持のためというか、intのままで扱っている箇所がいくつか残っている
* 航空攻撃の与ダメージ見積もりにおいて、噴式爆撃機の通常航空戦における攻撃力低下(/√2)を反映
* Calculator の各種判定を適切な？場所に移動
* 古い索敵式を非対応に(WIP)
* 補給時の消費資源計算を修正
* 艦隊：航空戦での威力について、噴式機の攻撃力が反映されていなかったのを修正
* 艦隊：昼戦空母カットインの威力補正に対応
* 艦隊：雷撃威力の中大破補正を修正
* 艦隊：夜戦駆逐カットイン(主雷電/魚見電)の威力補正に対応
* 艦船CSV出力に「改装段階」を追加

+++ ElectronicObserver/Data/Quest/ProgressDiscard.cs
@@ -59,7 +59,7 @@ namespace ElectronicObserver.Data.Quest
 			{
 				var eq = KCDatabase.Instance.Equipments[i];
 
-				if (Categories.Contains(eq.MasterEquipment.CategoryType))
+				if (Categories.Contains((int)eq.MasterEquipment.CategoryType))
 					Increment();
 			}
 		}
